backend/requirements.txt
# --- VIGIBYTE AI ENGINE DEPENDENCIES ---
# Purpose: Lists all Python libraries required to run the facial recognition backend.
# These versions are pinned to ensure stability across different deployment environments.

# Main web framework for building high-performance APIs
fastapi==0.115.0

# ASGI server implementation, used to run the FastAPI application
uvicorn[standard]==0.30.6

# The core AI library for face detection and 128-d encoding extraction (Built on dlib)
face-recognition==1.3.0

# Python Imaging Library (PIL fork) used for image opening and manipulation
Pillow==10.4.0

# Fundamental package for scientific computing; handles facial descriptors as numerical arrays
numpy==1.26.4

# Required by FastAPI to handle form data and file uploads (Base64/Multipart)
python-multipart==0.0.9